Working foreign repository and ref.
