+++ lib/elm_install.rb
@@ -14,8 +14,17 @@ require 'solve'
 
 require_relative './elm_install/version'
 require_relative './elm_install/utils'
-require_relative './elm_install/resolver'
 require_relative './elm_install/logger'
+require_relative './elm_install/ext'
+require_relative './elm_install/base'
+require_relative './elm_install/types'
+require_relative './elm_install/source'
+require_relative './elm_install/git_source'
+require_relative './elm_install/dependency'
+require_relative './elm_install/identifier'
+require_relative './elm_install/resolver'
+require_relative './elm_install/populator'
+require_relative './elm_install/installer'
 
 # The main module for the gem.
 module ElmInstall
