+++ lib/elm_install/base.rb
@@ -1,0 +1,6 @@
+module ElmInstall
+  class Base
+    include Contracts::Core
+    include Contracts::Builtin
+  end
+end
