+++ lib/elm_install/ext.rb
@@ -1,0 +1,13 @@
+module Semverse
+  class Version
+    def to_simple
+      "#{major}.#{minor}.#{patch}"
+    end
+
+    def self.try_new(version)
+      new version
+    rescue
+      nil
+    end
+  end
+end
