+++ lib/elm_install/installer.rb
@@ -1,0 +1,34 @@
+module ElmInstall
+  class Installer < Base
+    def initialize
+      @identifier = Identifier.new Dir.new(Dir.pwd)
+      @resolver = Resolver.new @identifier
+
+      puts "Resolving packages..."
+
+      @graph = @resolver.resolve
+
+      initial = @identifier.initial_dependencies.map do |dependency|
+        dependency.constraints.map do |constraint|
+          [dependency.name, constraint]
+        end
+      end
+      .flatten(1)
+
+      puts "Solving dependencies..."
+
+      results =
+        Solve
+          .it!(@graph, initial)
+          .map { |name, version|
+            dep = @resolver.dependencies[name]
+            dep.version = Semverse::Version.new(version)
+            dep
+          }
+
+      (Populator.new results).populate
+
+      puts "Packages configured successfully!"
+    end
+  end
+end
