+++ lib/elm_install/source.rb
@@ -1,0 +1,4 @@
+module ElmInstall
+  class Source < Base
+  end
+end
